+++ src/layout/sidebar.component.ts
@@ -1,9 +1,24 @@
-import { Component, HostBinding, Host, Optional, OnInit, Input, ElementRef, Output, EventEmitter, TemplateRef } from '@angular/core';
+import {
+    Component,
+    HostBinding,
+    Host,
+    Optional,
+    OnInit,
+    Input,
+    ElementRef,
+    Output,
+    EventEmitter,
+    TemplateRef,
+    OnDestroy
+} from '@angular/core';
 import { ThyLayoutComponent } from './layout.component';
 import { coerceBooleanProperty } from 'ngx-tethys/util';
 import { InputBoolean } from 'ngx-tethys/core';
 import { ThyResizeEvent } from 'ngx-tethys/resizable';
 import { isMacPlatform } from '@tethys/cdk/is';
+import { ThyHotkeyDispatcher } from '@tethys/cdk/hotkey';
+import { Subject, Subscription } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 
 const LG_WIDTH = 300;
 const SIDEBAR_DEFAULT_WIDTH = 240;
@@ -42,7 +57,6 @@ export type ThySidebarTheme = 'white' | 'light' | 'dark';
             class="sidebar-collapse"
             [ngClass]="{ 'collapse-visible': collapseVisible, 'collapse-hidden': collapseHidden }"
             (click)="toggleCollapse($event)"
-            [thyHotkey]="'Control+/,Meta+/'"
             [thyTooltip]="!thyTrigger && collapseTip"
         >
             <ng-template [ngTemplateOutlet]="thyTrigger || defaultTrigger"></ng-template>
@@ -52,7 +66,7 @@ export type ThySidebarTheme = 'white' | 'light' | 'dark';
         </div>
     `
 })
-export class ThySidebarComponent implements OnInit {
+export class ThySidebarComponent implements OnInit, OnDestroy {
     @HostBinding('class.thy-layout-sidebar') thyLayoutSidebarClass = true;
 
     @HostBinding('class.thy-layout-sidebar--clear-border-right') thyLayoutSidebarClearBorderRightClass = false;
@@ -103,7 +117,18 @@ export class ThySidebarComponent implements OnInit {
     @Output()
     thyDragWidthChange = new EventEmitter<number>();
 
-    @Input() @InputBoolean() thyCollapsible = false;
+    @Input() @InputBoolean() set thyCollapsible(collapsible: boolean) {
+        this.collapsible = collapsible;
+        if (this.collapsible) {
+            this.subscribeHotkeyEvent();
+        } else {
+            this.hotkeySubscription?.unsubscribe();
+        }
+    }
+
+    get thyCollapsible() {
+        return this.collapsible;
+    }
 
     @Input() @InputBoolean() set thyCollapsed(value: boolean) {
         this.isCollapsed = value;
@@ -136,6 +161,8 @@ export class ThySidebarComponent implements OnInit {
 
     collapseTip: string;
 
+    collapsible: boolean;
+
     isCollapsed = false;
 
     originWidth: number = SIDEBAR_DEFAULT_WIDTH;
@@ -146,13 +173,30 @@ export class ThySidebarComponent implements OnInit {
 
     isRemoveTransition: boolean;
 
-    constructor(@Optional() @Host() private thyLayoutComponent: ThyLayoutComponent, public elementRef: ElementRef) {}
+    private hotkeySubscription: Subscription;
+
+    constructor(
+        @Optional() @Host() private thyLayoutComponent: ThyLayoutComponent,
+        public elementRef: ElementRef,
+        private hotkeyDispatcher: ThyHotkeyDispatcher
+    ) {}
 
     ngOnInit() {
         if (this.thyLayoutComponent) {
             this.thyLayoutComponent.hasSidebar = true;
         }
         this.updateCollapseTip();
+    }
+
+    private subscribeHotkeyEvent() {
+        this.hotkeySubscription = this.hotkeyDispatcher.keydown(['Control+/', 'Meta+/']).subscribe(() => {
+            this.toggleCollapse();
+        });
+    }
+
+    private updateCollapseTip() {
+        this.collapseTip = this.thyCollapsed ? '展开' : '收起';
+        this.collapseTip = this.collapseTip + (isMacPlatform() ? `（⌘ + /)` : `（Ctrl + /)`);
     }
 
     resizeHandler({ width }: ThyResizeEvent) {
@@ -189,14 +233,13 @@ export class ThySidebarComponent implements OnInit {
         this.collapseVisible = type === 'enter' ? true : false;
     }
 
-    private updateCollapseTip() {
-        this.collapseTip = this.thyCollapsed ? '展开' : '收起';
-        this.collapseTip = this.collapseTip + (isMacPlatform() ? `（⌘ + /)` : `（Ctrl + /)`);
-    }
-
-    toggleCollapse(event: MouseEvent) {
+    toggleCollapse(event?: MouseEvent) {
         this.thyCollapsed = !this.thyCollapsed;
         setTimeout(() => this.updateCollapseTip(), 200);
         this.thyCollapsedChange.emit(this.isCollapsed);
+    }
+
+    ngOnDestroy(): void {
+        this.hotkeySubscription?.unsubscribe();
     }
 }