+++ src/message/index.ts
@@ -1,0 +1,7 @@
+export * from './module';
+export * from './message-container.component';
+export * from './message.component';
+export * from './message.service';
+export * from './message-queue.service';
+export * from './message.config';
+export * from './message-ref';